Add redux router state type
