+++ src/store/index.tsx
@@ -1,16 +1,9 @@
-import { composeWithDevTools } from '@redux-devtools/extension';
-import { createBrowserHistory } from 'history';
-import { applyMiddleware, legacy_createStore as createStore } from 'redux';
-import thunk from 'redux-thunk';
-import { combineReducers } from 'redux';
-import { createRouterMiddleware, createRouterReducer } from '@lagunovsky/redux-react-router';
-import rootReducer from './reducers';
+import { legacy_createStore as createStore } from 'redux';
 
-import { persistStore, persistReducer, createTransform } from 'redux-persist';
+import { createTransform, persistReducer, persistStore } from 'redux-persist';
+
+import { combinedReducer, devTools } from './configure';
 import storage from 'redux-persist/lib/storage';
-
-export const history = createBrowserHistory();
-const routerMiddleware = createRouterMiddleware(history);
 
 const persistConfig = {
   key: 'root',
@@ -26,22 +19,7 @@ const persistConfig = {
   blacklist: ['router', 'spinner', 'device', 'searching', 'options'],
 };
 
-const combinedReducer = combineReducers({
-  router: createRouterReducer(history),
-  ...rootReducer,
-});
-
 const persistedReducer = persistReducer(persistConfig, combinedReducer);
-
-const middleware = applyMiddleware(thunk, routerMiddleware);
-const devTools =
-  process.env.NODE_ENV === 'production'
-    ? middleware
-    : composeWithDevTools({
-        maxAge: 50,
-        actionSanitizer: (action: any) =>
-          action.type === 'SET_DATA' && action.payload ? { ...action, payload: '<<LARGE_PAYLOAD>>' } : action,
-      })(middleware);
 
 export default function configureStore() {
   const store = createStore(persistedReducer, devTools);
